Improved test coverage. Removed tests for Win32.
Currently all test cases that involve calling stacktrace are failing for 32-bit builds.
It appears that none of the test functions are showing up in the stacktrace at all,
probably because calls to Base.backtrace are returning different results. This seems to be
an issue in Julia, rather than StackTraces.jl, but I may be mistaken, as my ability to run
tests on 32-bit Windows is as limited as my time. I have removed these builds for now.